refactored feature property reading
