+++ bundles/mapping/mapwfs2/plugin/WfsVectorLayerPlugin/impl/AbstractLayerHandler.ol.js
@@ -1,7 +1,11 @@
+import { getFieldsAndPropsArrays } from '../util/props';
+const FEATURE_DATA_UPDATE_THROTTLE = 1000;
+
 export class AbstractLayerHandler {
     constructor (layerPlugin) {
         this.plugin = layerPlugin;
         this.layerIds = [];
+        this.throttledUpdates = new Map();
         this._log = Oskari.log('Oskari.mapping.mapmodule.AbstractLayerHandler');
     }
     /**
@@ -40,8 +44,54 @@ export class AbstractLayerHandler {
      * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
      * @param {ol/source} source
      */
-    updateLayerProperties (layer, source) {
-        this._log.debug('TODO: updateLayerProperties() not implemented on LayerHandler');
+    updateLayerProperties (layer, source = this._getLayerSource(layer)) {
+        if (this.throttledUpdates.has(layer.getId())) {
+            const throttledUpdate = this.throttledUpdates.get(layer.getId());
+            throttledUpdate();
+            return;
+        }
+        const update = () => this._updateLayerProperties(layer, source);
+        const throttledUpdate = Oskari.util.throttle(update, FEATURE_DATA_UPDATE_THROTTLE, { leading: false });
+        this.throttledUpdates.set(layer.getId(), throttledUpdate);
+        throttledUpdate();
+    }
+    _updateLayerProperties (layer, source) {
+        if (!layer.isVisible()) {
+            layer.setActiveFeatures([]);
+            layer.setFields([]);
+            this.plugin.notify('WFSPropertiesEvent', layer, layer.getLocales(), []);
+            return;
+        }
+        const { left, bottom, right, top } = this.plugin.getSandbox().getMap().getBbox();
+        const propsList = this._getFeaturePropsInExtent(source, [left, bottom, right, top]);
+        const { fields, properties } = getFieldsAndPropsArrays(propsList);
+        layer.setActiveFeatures(properties);
+        // Update fields and locales only if fields is not empty and it has changed
+        if (fields && fields.length > 0 && !Oskari.util.arraysEqual(layer.getFields(), fields)) {
+            layer.setFields(fields);
+            this.plugin.setLayerLocales(layer);
+        }
+        this.plugin.notify('WFSPropertiesEvent', layer, layer.getLocales(), fields);
+    }
+    _getFeaturePropsInExtent (source, extent) {
+        if (typeof source.getFeaturePropsInExtent === 'function') {
+            return source.getFeaturePropsInExtent(extent);
+        }
+        if (typeof source.getFeaturesInExtent === 'function') {
+            return source.getFeaturesInExtent(extent).map(ftr => ftr.getProperties());
+        }
+        return [];
+    }
+    /**
+     * Returns source corresponding to given layer
+     * @param {Oskari.mapframework.bundle.mapwfs2.domain.WFSLayer} layer
+     * @return {ol/source/VectorTile}
+     */
+    _getLayerSource (oskariLayer) {
+        const olLayers = this.plugin.getOLMapLayers(oskariLayer);
+        if (olLayers && olLayers.length > 0) {
+            return olLayers[0].getSource();
+        }
     }
     /**
      * @private @method _updateLayerStyle
