+++ bundles/mapping/mapwfs2/plugin/WfsVectorLayerPlugin/util/props.js
@@ -1,7 +1,7 @@
 export const WFS_ID_KEY = '_oid';
 export const WFS_FTR_ID_KEY = '__fid';
 
-const hiddenProps = new Set(['layer', WFS_ID_KEY]);
+const hiddenProps = new Set(['layer', 'geometry', WFS_ID_KEY]);
 
 function sortedFieldsFromProps (properties) {
     const fields = Object.keys(properties).filter(key => !hiddenProps.has(key));
